api/requirements.txt
Django==5.1.1
psycopg[binary]==3.2.1
pandas==2.2.2 
numpy==2.1.1
openpyxl==3.1.5
djangorestframework==3.15.2 
pydantic==2.8.2

django-cors-headers==4.4.0 